requirements.txt
asgiref==3.11.0
Django==4.2.27
django-cors-headers==4.9.0
djangorestframework==3.16.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.4
typing_extensions==4.15.0
reportlab==4.2.5
